Submitting stuff. But it crashes
Added PitDAO and PitReporsitory (Don't know if implementation is correct). Also changed AppDatabase to include the new pit stuff (Also don't know if implementation is correct). Right now when the submit button is pressed the app crashes. This happens with match scouting as well. Is this something which is fixed in newer branches?